Move AgentToolFactory and CallableActivation to graph.interfaces, as they are not actually interfaces used by RunContexts
