+++ neuro_san/internals/graph/interfaces/callable_activation.py
