+++ neuro_san/internals/interfaces/agent_tool_factory_provider.py
@@ -10,7 +10,7 @@
 #
 # END COPYRIGHT
 
-from neuro_san.internals.run_context.interfaces.agent_tool_factory import AgentToolFactory
+from neuro_san.internals.graph.interfaces.agent_tool_factory import AgentToolFactory
 
 
 class AgentToolFactoryProvider:
